Move branded navigation and split controllers to base
