+++ Packages/EurofurenceApplication/Sources/EurofurenceApplication/Content Router/Routes/News/ResetNewsAfterLogout.swift
@@ -1,3 +1,4 @@
+import ComponentBase
 import UIKit
 
 struct ResetNewsAfterLogout: NewsPresentation {
