--- Packages/EurofurenceApplication/Sources/EurofurenceApplication/Content Router/Wireframes/DetailPresentationContext.swift
@@ -1,4 +1,0 @@
-enum DetailPresentationContext {
-    case show
-    case replace
-}
