+++ Packages/EurofurenceApplication/Sources/EurofurenceApplication/Content Router/Wireframes/WindowContentWireframe.swift
@@ -1,3 +1,4 @@
+import ComponentBase
 import UIKit
 
 struct WindowContentWireframe: ContentWireframe {
