+++ Packages/EurofurenceComponents/Sources/ComponentBase/Assets.xcassets/Graphics/No Content Placeholder/No Content Placeholder.imageset/Contents.json
