+++ Packages/EurofurenceComponents/Sources/ComponentBase/Assets.xcassets/Graphics/No Content Placeholder/No Content Placeholder.imageset/eurofurence_logo_vector_full_2017 white crop.pdf
