+++ Packages/EurofurenceComponents/Sources/ComponentBase/Branding/BrandedNavigationController.swift
@@ -1,0 +1,13 @@
+import UIKit
+
+public class BrandedNavigationController: UINavigationController {
+    
+    override public func viewDidLoad() {
+        super.viewDidLoad()
+        
+        if #available(iOS 13.0, *) {
+            navigationBar.prefersLargeTitles = true
+        }
+    }
+    
+}
