+++ Packages/EurofurenceComponents/Sources/ComponentBase/No Content Placeholder/NoContentPlaceholderViewController.swift
@@ -1,15 +1,15 @@
 import UIKit
 
-class NoContentPlaceholderViewController: UIViewController {
+public class NoContentPlaceholderViewController: UIViewController {
     
-    static func fromStoryboard() -> NoContentPlaceholderViewController {
+    public static func fromStoryboard() -> NoContentPlaceholderViewController {
         let storyboard = UIStoryboard(name: "NoContentPlaceholderViewController", bundle: .module)
         return storyboard.instantiate(NoContentPlaceholderViewController.self)
     }
     
     @IBOutlet private weak var placeholderImageView: UIImageView!
     
-    override func viewDidLoad() {
+    override public func viewDidLoad() {
         super.viewDidLoad()
         placeholderImageView.tintColor = .placeholder
     }
